reorganize s1_disp_stack into two workflows
