+++ src/dolphin/__init__.py
@@ -4,3 +4,4 @@ from dolphin._version import version as __version__
 
 from ._log import *
 from ._show_versions import *
+from .io import *
